Fix: Reduce Code
