+++ src/components/LoginCard/CompanyForms/index.tsx
@@ -8,6 +8,7 @@ import { EmailIcon } from '../../EmailIcon';
 import { PasswordIcon } from '../../PasswordIcon';
 import { PopUpRegisterSucess } from '../PopUpRegisterSuccess';
 import { normalizeCnpjNumber } from '../MaskedInput/MaskedInput';
+import PopupHandler from '../PopUpRegisterSuccess/CloseModalEsc';
 import {
     schemaCompanyLoginForm,
     schemaCompanyRegisterForm,
@@ -74,37 +75,8 @@ export const CompanyForms = (props: any): JSX.Element => {
         resolver: yupResolver(getFormValidation),
     });
 
-    const handleKeyDown = useCallback(
-        (event: KeyboardEvent) => {
-            if (event.key === 'Escape') {
-                setPopup(false);
-            }
-        },
-        [setPopup],
-    );
-    useEffect(() => {
-        const handleOutsideClick = (event: MouseEvent) => {
-            const target = event.target as HTMLElement;
-            if (!target.closest(Popup)) {
-                setPopup(false);
-            }
-        };
-        document.addEventListener('mousedown', handleOutsideClick);
-
-        return () => {
-            document.removeEventListener('mousedown', handleOutsideClick);
-        };
-    }, [setPopup]);
-
-    useEffect(() => {
-        document.addEventListener('keydown', handleKeyDown);
-        return () => {
-            document.removeEventListener('keydown', handleKeyDown);
-        };
-    }, [handleKeyDown]);
-
     const characters = /^(?=.{8,20}$).*$/;
-    const letters = /^(?=.*[a-z])(?=.*[A-Z])\w+$/;
+    const letters = /^(?=.*[a-zA-Z]).*$/;
     const number = /^(?=.*\d).*$/;
     const specialCharacters = /^(?=.*\W).*$/;
 
@@ -445,6 +417,7 @@ export const CompanyForms = (props: any): JSX.Element => {
                     close={closePopup}
                 />
             ) : null}
+            <PopupHandler setPopup={setPopup} Popup={Popup} />
         </>
     );
 };
